// openk9-3.1.x/js-packages/admin-ui/src/components/Navigation/types.ts
/*
* Copyright (c) 2020-present SMC Treviso s.r.l. All rights reserved.
*
* This program is free software: you can redistribute it and/or modify
* it under the terms of the GNU Affero General Public License as published by
* the Free Software Foundation, either version 3 of the License, or
* (at your option) any later version.
*
* This program is distributed in the hope that it will be useful,
* but WITHOUT ANY WARRANTY; without even the implied warranty of
* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
* GNU Affero General Public License for more details.
*
* You should have received a copy of the GNU Affero General Public License
* along with this program.  If not, see <http://www.gnu.org/licenses/>.
*/
import { NamePath } from "../sideNavigationContext";

export type MenuItem = {
  label: string;
  path?: string;
  IsChildren: boolean;
  value: NamePath;
  isGroup?: boolean;
  children?: MenuItem[];
}; 